refactor so that only env.js access process.env. Defaults are set in this file
